feature36: Product Server 숙박업체 재고 조회 Service - BookingStoreImpl -> AccommodationService 의존성 제거 - BookingService -> AccommodationStore 추가
