+++ expedia-server/booking-server/src/main/java/org/clonecoder/bookingserver/domain/booking/BookingStore.java
@@ -11,5 +11,4 @@ import java.util.List;
 public interface BookingStore {
     Booking saveBooking(Booking booking);
     void saveBookingGuests(List<BookingGuests> bookingGuests);
-    void accommodationRoomStockDecrease(Long accommodationRoomId);
 }
