+++ expedia-server/booking-server/src/main/java/org/clonecoder/bookingserver/infrastructure/BookingStoreImpl.java
@@ -4,9 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.clonecoder.bookingserver.domain.Booking;
 import org.clonecoder.bookingserver.domain.BookingGuests;
 import org.clonecoder.bookingserver.domain.booking.BookingStore;
-import org.clonecoder.productserver.domain.accommodation.AccommodationService;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -15,7 +13,6 @@ import java.util.List;
 public class BookingStoreImpl implements BookingStore {
     private final BookingRepository bookingRepository;
     private final BookingGuestsRepository bookingGuestsRepository;
-    private final AccommodationService accommodationService;
 
     /**
      * 예약 저장
@@ -31,13 +28,5 @@ public class BookingStoreImpl implements BookingStore {
      */
     public void saveBookingGuests(List<BookingGuests> bookingGuests) {
         bookingGuestsRepository.saveAll(bookingGuests);
-    }
-
-    /**
-     * 예약건에 대한 재고 설정
-     * @param accommodationRoomId
-     */
-    public void accommodationRoomStockDecrease(Long accommodationRoomId) {
-        accommodationService.stockDecrease(accommodationRoomId);
     }
 }
