+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/domain/accommodation/AccommodationService.java
@@ -11,16 +11,5 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Slf4j
 public class AccommodationService {
-    private final AccommodationStore accommodationStore;
 
-    @RedissonLock(key = "accommodationRoomId")
-    public void stockDecrease(Long accommodationRoomId) {
-        log.info("Thread Name : " + Thread.currentThread().getName());
-
-        AccommodationRoom accommodationRoom = accommodationStore.getAccommodationRoom(accommodationRoomId);
-        log.info("stock() : " + accommodationRoom.getStock());
-
-        accommodationRoom.stockDecrease();
-        log.info("decrease after stock() : " + accommodationRoom.getStock());
-    }
 }
