+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/domain/accommodation/AccommodationStore.java
@@ -3,5 +3,5 @@ package org.clonecoder.productserver.domain.accommodation;
 import org.clonecoder.productserver.domain.AccommodationRoom;
 
 public interface AccommodationStore {
-    AccommodationRoom getAccommodationRoom(Long accommodationRoomId);
+    void stockDecrease(Long accommodationRoomId);
 }
