+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/infrastructure/AccommodationStoreImpl.java
@@ -1,20 +1,29 @@
 package org.clonecoder.productserver.infrastructure;
 
+import lombok.extern.slf4j.Slf4j;
+import org.clonecoder.productserver.common.aop.RedissonLock;
 import org.clonecoder.productserver.domain.AccommodationRoom;
 import org.clonecoder.productserver.domain.accommodation.AccommodationStore;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class AccommodationStoreImpl implements AccommodationStore {
     private final AccommodationRoomRepository accommodationRoomRepository;
 
-    @Override
-    public AccommodationRoom getAccommodationRoom(Long accommodationRoomId) {
-        return accommodationRoomRepository.findById(accommodationRoomId).get();
+    @RedissonLock(key = "accommodationRoomId")
+    public void stockDecrease(Long accommodationRoomId) {
+        log.info("Thread Name : " + Thread.currentThread().getName());
+
+        AccommodationRoom accommodationRoom = accommodationRoomRepository.findById(accommodationRoomId).get();
+        log.info("stock() : " + accommodationRoom.getStock());
+
+        accommodationRoom.stockDecrease();
+
+        log.info("decrease after stock() : " + accommodationRoom.getStock());
     }
 }
